Models for database tables and api responses :label:
